fix: CMDT auf Integer-Protokoll umstellen

+++ 5_System/src/Parser/CommandParser.h
@@ -10,16 +10,22 @@ enum CmdType : uint8_t
 {
     CMD_NONE = 0,
     CMD_TIME = 1,
-    CMD_PATH = 2,   // später
+    CMD_PATH = 2,   // spaeter
 };
 
 struct ParsedCommand
 {
     CmdType type;
-    float vx;
-    float vy;
-    float wz;
-    float param;     // CMD_TIME: Sekunden, CMD_PATH: cm
+
+    // Integer-Protokoll:
+    // vx    = Geschwindigkeit in cm/s
+    // vy    = Geschwindigkeit in cm/s
+    // wz    = Winkelgeschwindigkeit in Grad/s
+    // param = Dauer in Sekunden bei CMDT
+    int16_t  vx;
+    int16_t  vy;
+    int16_t  wz;
+    uint16_t param;
 };
 
 class CommandParser
